ferramentas_catalogo/requirements.txt
gradio>=4.0.0
rembg[gpu]>=2.0.57
Pillow>=10.0.0
